@@ -1,11 +1,11 @@
-module github.com/dairaga/js
+module github.com/dairaga/js/v2
 
-go 1.12
+go 1.18
+
+require github.com/stretchr/testify v1.7.1
 
 require (
-	github.com/agnivade/wasmbrowsertest v0.3.0 // indirect
-	github.com/dairaga/webkit v0.0.0-20190820144517-bd7ab10bed51
-	github.com/golang/protobuf v1.3.2
-	github.com/shurcooL/go v0.0.0-20190704215121-7189cc372560 // indirect
-	github.com/shurcooL/go-goon v0.0.0-20170922171312-37c2f522c041 // indirect
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
