@@ -1,331 +1,45 @@
+//go:build js && wasm
+
+// Package xhr is a wrapper of Javascript XMLHttpRequest (XHR), and provides a simple way to make a JSON AJAX.
+//
+// See https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest.
 package xhr
 
 import (
-	"encoding/json"
 	"fmt"
-	"reflect"
-	"strings"
-
-	"github.com/dairaga/js"
 )
 
-// State xhr state
-type State int
+type Error int
 
-// ReadyState ...
 const (
-	Unsent          State = 0
-	Opened          State = 1
-	HeadersReceived State = 2
-	Loading         State = 3
-	Done            State = 4
+	ErrAbort    Error = -1 // Request aborted.
+	ErrFailed   Error = -2 // Request failed.
+	ErrTimeout  Error = -3 // Request timeout.
+	ErrReleased Error = -4 // Request released.
+
+	GET    = "GET"    // HTTP GET method.
+	POST   = "POST"   // HTTP POST method.
+	PUT    = "PUT"    // HTTP PUT method.
+	DELETE = "DELETE" // HTTP DELETE method.
+	PATCH  = "PATCH"  // HTTP PATCH method.
 )
 
-// content types
-const (
-	TextType   string = "text/plain"
-	JSONType   string = "application/json"
-	StreamType string = "application/octet-stream"
-)
+// -----------------------------------------------------------------------------
 
-func (s State) String() string {
-	switch s {
-	case Unsent:
-		return "Unsent"
-	case Opened:
-		return "Opened"
-	case HeadersReceived:
-		return "Headers Received"
-	case Loading:
-		return "Loading"
-	case Done:
-		return "Done"
+func (e Error) Error() string {
+	switch e {
+	case ErrAbort:
+		return "abort"
+	case ErrFailed:
+		return "error"
+	case ErrTimeout:
+		return "timeout"
+	case ErrReleased:
+		return "released"
 	default:
-		return fmt.Sprintf("Unknown(%d)", s)
-	}
-}
-
-var xhrConstructor = js.Global().Get("XMLHttpRequest")
-
-//-----------------------------------------------------------------------------
-
-// Request ...
-type Request struct {
-	method  string
-	url     string
-	headers map[string]string
-	body    []byte
-}
-
-// NewRequest returns a request.
-func NewRequest(method, url string, body []byte) *Request {
-	return &Request{
-		method: method,
-		url:    url,
-		body:   body,
-	}
-}
-
-// SetHeader set request header.
-func (req *Request) SetHeader(key, value string) *Request {
-
-	if req.headers == nil {
-		req.headers = make(map[string]string)
-	}
-
-	req.headers[key] = value
-	return req
-}
-
-//-----------------------------------------------------------------------------
-
-// Response ...
-type Response struct {
-	code    int
-	headers map[string]string
-	body    []byte
-}
-
-func (resp *Response) fill(xhr js.Value) *Response {
-	resp.code = xhr.Get("status").Int()
-
-	if xhr.Get("readyState").Int() == int(Done) {
-		resp.body = js.Bytes(xhr.Get("response"))
-	}
-
-	headers := xhr.Call("getAllResponseHeaders").String()
-
-	tmp := strings.Split(headers, "\r\n")
-	if size := len(tmp); size > 0 {
-		resp.headers = make(map[string]string, size)
-		for _, x := range tmp {
-			parts := strings.Split(x, ": ")
-			resp.headers[parts[0]] = strings.Join(parts[1:], ": ")
+		if e >= StatusBadRequest {
+			return fmt.Sprintf("code: %d, status: %s", e, StatusText(int(e)))
 		}
+		return fmt.Sprintf("unknown error (%d)", e)
 	}
-	return resp
-}
-
-// Code ...
-func (resp *Response) Code() int {
-	return resp.code
-}
-
-// Body ...
-func (resp *Response) Body() []byte {
-	return resp.body
-}
-
-// GetHeader ...
-func (resp *Response) GetHeader(key string) string {
-	if resp.headers != nil {
-		return resp.headers[key]
-	}
-	return ""
-}
-
-func (resp *Response) String() string {
-	codestr := fmt.Sprintf(`"code": %d`, resp.code)
-	bodystr := fmt.Sprintf(`"body": %s`, string(resp.body))
-	headers := ""
-	for k, v := range resp.headers {
-		headers += fmt.Sprintf(`%q: %q,`, k, v)
-	}
-	return fmt.Sprintf(`{%s, headers: {%s}}, %s`, codestr, headers[:len(headers)-1], bodystr)
-}
-
-//-----------------------------------------------------------------------------
-
-// Client ...
-type Client struct {
-	ref        js.Value
-	resp       *Response
-	lastErr    error
-	onDone     func(*Response)
-	onFail     func(*Response)
-	onAlways   func(*Response)
-	onProgress func(int, int, bool)
-}
-
-// NewClient returns Client
-func NewClient() *Client {
-	cli := &Client{
-		ref: xhrConstructor.New(),
-	}
-
-	cli.ref.Set("responseType", "arraybuffer")
-
-	cli.ref.Call("addEventListener", "load", js.FuncOf(func(_ js.Value, args []js.Value) interface{} {
-		if cli.onDone != nil {
-			if cli.resp == nil {
-				cli.resp = new(Response)
-			}
-			cli.resp.fill(cli.ref)
-			if cli.resp.code >= 200 && cli.resp.code < 300 {
-				cli.onDone(cli.resp)
-			} else {
-				if cli.onFail != nil {
-					cli.onFail(cli.resp)
-				}
-			}
-		}
-		return nil
-	}))
-
-	cli.ref.Call("addEventListener", "error", js.FuncOf(func(_ js.Value, args []js.Value) interface{} {
-		if cli.onFail != nil {
-			if cli.resp == nil {
-				cli.resp = new(Response)
-			}
-			cli.onFail(cli.resp.fill(cli.ref))
-		}
-		return nil
-	}))
-
-	cli.ref.Call("addEventListener", "loadend", js.FuncOf(func(_ js.Value, args []js.Value) interface{} {
-		if cli.onAlways != nil {
-			if cli.resp == nil {
-				cli.resp = new(Response)
-				cli.resp.fill(cli.ref)
-			}
-			cli.onAlways(cli.resp)
-		}
-		return nil
-	}))
-
-	cli.ref.Call("addEventListener", "progress", js.FuncOf(func(_ js.Value, args []js.Value) interface{} {
-		if cli.onProgress != nil {
-			cli.onProgress(args[0].Get("loaded").Int(), args[0].Get("total").Int(), args[0].Get("lengthComputable").Bool())
-		}
-		return nil
-	}))
-
-	return cli
-}
-
-// JSValue ...
-func (cli *Client) JSValue() js.Value {
-	return cli.ref
-}
-
-// LastErr ...
-func (cli *Client) LastErr() error {
-	return cli.lastErr
-}
-
-// ReadyState https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/readyState
-func (cli *Client) ReadyState() State {
-	return State(cli.ref.Get("readyState").Int())
-}
-
-// Do ...
-func (cli *Client) Do(req *Request) *Client {
-	cli.ref.Call("open", req.method, req.url, true)
-	if len(req.headers) > 0 {
-		for k, v := range req.headers {
-			cli.ref.Call("setRequestHeader", k, v)
-		}
-	}
-
-	if len(req.body) > 0 {
-		//arr := js.TypedArrayOf(req.body)
-		arr := js.ToJSUint8Array(req.body)
-		cli.ref.Call("send", arr)
-		//arr.Release()
-	} else {
-		cli.ref.Call("send")
-	}
-	return cli
-}
-
-// Done ...
-func (cli *Client) Done(cb func(*Response)) *Client {
-	cli.onDone = cb
-	return cli
-}
-
-// Fail ...
-func (cli *Client) Fail(cb func(*Response)) *Client {
-	cli.onFail = cb
-	return cli
-}
-
-// Always ...
-func (cli *Client) Always(cb func(*Response)) *Client {
-	cli.onAlways = cb
-	return cli
-}
-
-// Progress ...
-func (cli *Client) Progress(cb func(int, int, bool)) *Client {
-	cli.onProgress = cb
-	return cli
-}
-
-//-----------------------------------------------------------------------------
-
-func handle(data interface{}) (string, []byte, error) {
-	if data == nil {
-		return "", nil, nil
-	}
-
-	typ := reflect.TypeOf(data)
-	if typ.Kind() == reflect.Struct || (typ.Kind() == reflect.Ptr && typ.Elem().Kind() == reflect.Struct) {
-		dataBytes, err := json.Marshal(data)
-		if err != nil {
-			return "", nil, err
-		}
-
-		return JSONType, dataBytes, nil
-	}
-
-	switch v := data.(type) {
-	case string:
-		return TextType, []byte(v), nil
-	case []byte:
-		return StreamType, v, nil
-	}
-
-	return "", nil, fmt.Errorf("unknown type: %v", typ)
-}
-
-func do(method, url string, data []interface{}) *Client {
-	cli := NewClient()
-	var x interface{}
-	if len(data) > 0 {
-		x = data[0]
-	}
-	ctype, dataBytes, err := handle(x)
-	if err != nil {
-		cli.lastErr = err
-		return cli
-	}
-
-	req := NewRequest(method, url, dataBytes)
-	if ctype != "" {
-		req.SetHeader("Content-Type", ctype)
-	}
-
-	cli.Do(req)
-	return cli
-}
-
-// Get ...
-func Get(url string, data ...interface{}) *Client {
-	return do("GET", url, data)
-}
-
-// Post ...
-func Post(url string, data ...interface{}) *Client {
-	return do("POST", url, data)
-}
-
-// Put ...
-func Put(url string, data ...interface{}) *Client {
-	return do("PUT", url, data)
-}
-
-// Delete ...
-func Delete(url string, data ...interface{}) *Client {
-	return do("DELETE", url, data)
 }
